backend/requirements.txt
click==7.1.2
distro==1.5.0
fastapi==0.54.1
h11==0.9.0
httptools==0.1.1
numpy==1.18.4
pandas==1.0.3
pkg-resources==0.0.0
pydantic==1.5.1
python-dateutil==2.8.1
python-multipart==0.0.5
pytz==2020.1
six==1.14.0
starlette==0.13.2
tabula-py==2.1.0
uvicorn==0.11.5
uvloop==0.14.0
websockets==8.1
